breaking: remove implicit Hs from RDKit SMILES
Fix #2087.
