+++ tests/test.json
@@ -40,7 +40,7 @@
         ]
       ]
     },
-    "reaction_sha256": "5eb802a5a4f5c537ea556528a7db9402ab401f9cbd650e75b2a03e809ab9c8ad"
+    "reaction_sha256": "12a2c53d81db733405d24dfc67dd390c85195c4011cef53305629e8083a5d2f4"
   },
   {
     "rngparams": {
@@ -64,7 +64,8 @@
       "speciescenter": "[H][H]"
     },
     "reaction_sha256": [
-      "3ad36fbbbe2f52f4533b42e19e9efa393129d505860f00efb544aac7820caac7"
+      "7b26f8b66b9134a726e1aadc4adf200f42f5921692d30a1c7867fcf437b1fb88",
+      "1f3a88d3281b7627bc9bed1a533d7b08f1edd4d7811291d4d07cc5bbf7ac9bd4"
     ]
   },
   {
@@ -88,8 +89,10 @@
       "miso": 1
     },
     "reaction_sha256": [
-      "ec826635846cbe4d70a2454056ae973aaaf2c7bb760514017ebd6c804a85b6c2",
-      "2b1f9276faeabd4dbbd3349a9ddd5ffbcb64ebc11dcf3f16d661c34372f36dca"
+      "c7d2d7b905d247aaeba36b0c272506ab287c557c5365b7090bd9a04a45e1fc48",
+      "3d99317c870758bd9d33e29d474d9496643b74bd7b4b07fa1e31afff33dbcd07",
+      "88e88f1fd4e83f37f560b6786cee0feadf1139bfb7de8551b2946a11e4b767f9",
+      "e1f445eab53f7208ac9b176a09ac79360999957edbf2e5a9d9e8990bc20b91d7"
     ]
   },
   {
